requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
Pillow==9.5.0
torch==2.0.0
pytorch-lightning==1.8.3
torchvision==0.15.1
efficientnet_pytorch==0.7.1